homework/stage02_tooling-setup_slides-outline/requirements.txt
python-dotenv
numpy
jupyter
